+++ src/QsCompiler/QirGeneration/Subtransformations/NamespaceTransformation.cs
